requirements.txt
django
psycopg2
djangorestframework
markdown 
django-filter
libgen-api